scripts/mimamori-capture-screen/requirements.txt
python-dotenv
opencv-python == 4.6.0.66
requests